fixed probably most bugs with systable row-data conversions

+++ src/backend/internal/analyser/seganalyser/typanlr/future_type_builder.go
@@ -24,7 +24,7 @@ func createFTFromConst(expr node2.Const) FutureType {
 		}
 	case rawtype.TypeStr:
 		args := rawtype.Args{
-			Length: uint32(len(expr.Val)),
+			Length: len(expr.Val),
 			UTF8:   !rawtype.StrUtils.IsASCII(expr.Val),
 			VarLen: true,
 		}
